refactor(config): config.go（970行）を責務別ファイルへ分割（API変更なし）
- assets.go: JingleEntry/SEEntry/BGMEntry/AssetsConfig + validate/load helpers
- cache.go: CacheConfig + constants
- characters.go: CharacterConfig
- episode_spec.go: EpisodeSpec/EpisodeCondition/CastConfig + validate群
- llm.go: LLMConfig 関連型・定数
- load.go: LoadConfig/LoadEpisodeSpec 等のロード関数
- program.go: ProgramConfig/CornerConfig/SourceConfig/AudioRef/FeedEntry + constants
- slack.go: SlackConfig
- voicevox.go: VoicevoxPresets/VoicevoxConfig + presets vars
- config.go: Config struct のみに縮小
- docs/architecture/architecture.md §8 から T7 違反行を削除
